+++ lib/elixir/lib/behaviour.ex
@@ -118,10 +118,6 @@ defmodule Behaviour do
     end
   end
 
-  defp ensure_not_default({ ://, _, [_, _] }) do
-    raise ArgumentError, message: "default arguments // not supported in defcallback/defmacrocallback"
-  end
-
   defp ensure_not_default({ :\\, _, [_, _] }) do
     raise ArgumentError, message: "default arguments \\\\ not supported in defcallback/defmacrocallback"
   end