+++ lib/elixir/test/erlang/tokenizer_test.erl
@@ -123,9 +123,6 @@ empty_string_test() ->
   [{bin_string,1,[<<>>]}] = tokenize("\"\""),
   [{list_string,1,[<<>>]}] = tokenize("''").
 
-default_test() ->
-  [{identifier,1,x},{in_match_op,1,'//'},{number,1,1}] = tokenize("x // 1").
-
 addadd_test() ->
   [{identifier,1,x},{two_op,1,'++'},{identifier,1,y}] = tokenize("x ++ y").